+++ crates/stackable-versioned-macros/fixtures/inputs/default/generics_defaults.rs
@@ -1,9 +1,0 @@
-#[versioned(version(name = "v1alpha1"), version(name = "v1"))]
-// ---
-pub struct Foo<T = String>
-where
-    T: Default,
-{
-    bar: T,
-    baz: u8,
-}