+++ crates/stackable-versioned-macros/tests/inputs/default/pass/module_preserve.rs
@@ -1,0 +1,27 @@
+use stackable_versioned::versioned;
+// ---
+#[versioned(
+    version(name = "v1alpha1"),
+    version(name = "v1"),
+    version(name = "v2alpha1"),
+    options(preserve_module)
+)]
+// ---
+pub(crate) mod versioned {
+    pub struct Foo {
+        bar: usize,
+
+        #[versioned(added(since = "v1"))]
+        baz: bool,
+
+        #[versioned(deprecated(since = "v2alpha1"))]
+        deprecated_foo: String,
+    }
+
+    #[versioned]
+    pub struct Bar {
+        baz: String,
+    }
+}
+// ---
+fn main() {}