+++ crates/stackable-versioned-macros/tests/trybuild.rs
@@ -15,31 +15,65 @@
 // syntax highlighting and code completion. You will need to comment them out
 // again but before running tests, otherwise compilation will fail (as expected).
 #[allow(dead_code)]
-mod default {
-    // mod fail {
-    //     mod deprecate;
-    //     mod skip_from_all;
-    //     mod skip_from_version;
-    // }
+mod inputs {
+    mod default {
+        mod pass {
+            // mod attribute_enum;
+            // mod attribute_struct;
+            // mod basic_struct;
+            // mod convert_with;
+            // mod deprecate_enum;
+            // mod deprecate_struct;
+            // mod enum_data_simple;
+            // mod generics_defaults;
+            // mod generics_module;
+            // mod generics_struct;
+            // mod module;
+            // mod module_preserve;
+            // mod rename;
+            // mod skip_from_for_version;
+            // mod skip_from_module;
+            // mod skip_from_module_for_version;
+            // mod submodule;
+        }
+        mod fail {
+            // mod changed;
+            // mod deprecate;
+            // mod skip_from_all;
+            // mod skip_from_version;
+            // mod submodule_invalid_name;
+            // mod submodule_use_statement;
+        }
+    }
+
+    #[cfg(feature = "k8s")]
+    mod k8s {
+        mod pass {
+            // mod basic;
+            // mod crate_overrides;
+            // mod module;
+            // mod module_preserve;
+            // mod renamed_kind;
+            // mod shortnames;
+            // mod skip;
+        }
+
+        mod fail {
+            // mod crd;
+        }
+    }
 }
 
 #[test]
-fn default_macros() {
+fn default() {
     let t = trybuild::TestCases::new();
-    t.compile_fail("tests/default/fail/*.rs");
-}
-
-#[cfg(feature = "k8s")]
-#[allow(dead_code)]
-mod k8s {
-    // mod fail {
-    //     mod crd;
-    // }
+    t.compile_fail("tests/inputs/default/fail/*.rs");
+    t.pass("tests/inputs/default/pass/*.rs");
 }
 
 #[cfg(feature = "k8s")]
 #[test]
-fn k8s_macros() {
+fn k8s() {
     let t = trybuild::TestCases::new();
-    t.compile_fail("tests/k8s/fail/*.rs");
+    t.compile_fail("tests/inputs/k8s/fail/*.rs");
 }